Use longs instead of ints to store pointers in NIO buffers.
This allows the code to be used on 64-bit VMs.

+++ luni/src/main/java/java/nio/DirectByteBuffer.java
@@ -37,7 +37,7 @@ class DirectByteBuffer extends MappedByteBuffer {
       throw new IllegalArgumentException("capacity + offset > baseSize");
     }
 
-    this.effectiveDirectAddress = block.toInt() + offset;
+    this.effectiveDirectAddress = block.toLong() + offset;
 
     this.offset = offset;
     this.isReadOnly = isReadOnly;
